// packages/google-pay/interfaces.ts
export interface PaymentCancelledEventData extends EventData {
	eventName: string;
	object: any;
}

export interface PaymentErrorEventData extends EventData {
	eventName: string;
	object: any;
	data: {
		statusCode: number;
	};
}

export interface PaymentSuccessEventData extends EventData {
	eventName: string;
	object: any;
	data: {
		/**
		 * Data about the selected payment method.
		 */
		paymentMethodData: {
			/**
			 * PaymentMethod type selected in the Google Pay payment sheet.
			 */
			type: string;

			/**
			 * User-facing message to describe the payment method that funds this transaction.
			 */
			description: string;

			/**
			 * The value of this property depends on the payment method type returned. For CARD, see CardInfo.
			 */
			info: {
				/**
				 * The details about the card. This value is commonly the last four digits of the selected payment account number.
				 */
				cardDetails: string;

				/**
				 * This object provides information about the validation performed on the returned payment data if assuranceDetailsRequired is set to true in the CardParameters.
				 */
				assuranceDetails: {
					/**
					 * If true, indicates that Cardholder possession validation has been performed on returned payment credential.
					 */
					accountVerified: boolean;

					/**
					 * If true, indicates that identification and verifications (ID&V) was performed on the returned payment credential.
					 * If false, the same risk-based authentication can be performed as you would for card transactions. This risk-based authentication can include, but not limited to, step-up with 3D Secure protocol if applicable.
					 */
					cardHolderAuthenticated: boolean;
				};

				/**
				 * The payment card network of the selected payment. Returned values match the format of allowedCardNetworks in CardParameters.
				 * This card network value should not be displayed to the buyer. It's used when the details of a buyer's card are needed. For example, if customer support needs this value to identify the card a buyer used for their transaction. For a user-visible description, use the description property of PaymentMethodData instead.
				 */
				cardNetwork: string;

				/**
				 * The billing address associated with the provided payment method, if billingAddressRequired is set to true in CardParameters.
				 */
				billingAddress?: Address;
			};

			/**
			 * Payment tokenization data for the selected payment method.
			 */
			tokenizationData: {
				/**
				 * 	The type of tokenization to be applied to the selected payment method. This value matches the type set in PaymentMethodTokenizationSpecification.
				 */
				type: string;

				/**
				 * The generated payment method token.
				 * PAYMENT_GATEWAY: JSON object string that contains a chargeable token issued by your gateway.
				 * DIRECT: protocolVersion, signature, and a signedMessage for decryption. See Payment method token structure for more information.
				 */
				token: {
					/**
					 * Identifies the encryption or signing scheme under which the message is created. It allows the protocol to evolve over time, if needed.
					 */
					protocolVersion: string;

					/**
					 * Verifies that the message came from Google. It's base64-encoded, and created with ECDSA by the intermediate signing key.
					 */
					signature: string;

					/**
					 * A JSON object serialized as a string that contains the encryptedMessage, ephemeralPublicKey, and tag. It's serialized to simplify the signature verification process.
					 */
					signedMessage: string;
				};
			};
		};

		/**
		 * Email address, if emailRequired is set to true in the PaymentDataRequest. If another request has the property set to true there's no effect.
		 */
		email: string;

		/**
		 * Shipping address, if shippingAddressRequired is set to true in the PaymentDataRequest.
		 */
		shippingAddress: Address;
	};
}

export interface Address {
	/**
	 * The full name of the addressee.
	 */
	name;

	/**
	 * The postal or ZIP code.
	 */
	postalCode;

	/**
	 * ISO 3166-1 alpha-2 country code.
	 */
	countryCode;

	/**
	 * A telephone number, if phoneNumberRequired is set to true in the PaymentDataRequest.
	 */
	phoneNumber;

	/**
	 * The first line of the address.
	 */
	address1;

	/**
	 * The second line of the address.
	 */
	address2;

	/**
	 * The third line of the address.
	 */
	address3;

	/**
	 * City, town, neighborhood, or suburb.
	 */
	locality;

	/**
	 * A country subdivision, such as a state or province.
	 */
	administrativeArea;

	/**
	 * The sorting code.
	 */
	sortingCode;
}